Engadget: Deals
Get up to $700 off a new Microsoft Copilot+ PC before Windows 10 support ends

Deals at Best Buy include a souped-up Surface laptop for $1,100.

Lawrence Bonk

Contributing Reporter

Lawrence BonkContributing Reporter

Mon, October 13, 2025 at 4:57 PM UTC

Microsoft

Best Buy is holding a sitewide sale on Microsoft Copilot+ PCs with steep discounts up to $700 off. This is fantastic timing, given that Microsoft ends official support for Windows 10 this week. If you're thinking about upgrading, now is likely the time.

One cool deal here is for a souped-up version of the Microsoft Surface 7 laptop. This one ships with a Snapdragon X Plus processor, 32GB of RAM and 1TB of internal storage. It also includes a 13.8-inch touchscreen and a battery that gets 20 hours per charge. The laptop ships with Windows 11 Home, which should be relatively futureproof for the next several years. It costs just $1,100, which is a nice discount of $700.

Microsoft

Microsoft Surface 7 Laptop (32GB/1TB)

$1,100$1,800Save $700

This includes a Snapdragon X Plus CPU and a touchscreen.

$1,100 at Best Buy

This model can be outfitted with even more bells and whistles. There's one with a more powerful Snapdragon X Elite processor and a 15-inch touchscreen. That one costs $1,500, which is a discount of $600.

Advertisement

Advertisement

Other deals include the HP Omen gaming laptop with an AMD Ryzen AI 9 HX 375 and NVIDIA GeForce RTX 5070 Ti for $2,000 instead of $2,380 and the MSI Codex gaming desktop with an NVIDIA GeForce RTX 5070 GPU for $1,460 instead of $1,600. There are plenty of other discounted laptops, desktops and accessories on the main sale page.

Microsoft will officially end support of Windows 10 on October 14, but it may not be an immediate death knell. There is a way for users to get an additional 12 months of support via the company's Extended Security Updates plan. Many Windows 10 computers can also easily handle Windows 11, and a software update would be much cheaper than buying a new computer.

Follow @EngadgetDeals on X for the latest tech deals and buying advice.

Advertisement

About our ads