Techcrunch:
Apple announced Tuesday that it had signed contracts for 650 megawatts of renewable power in Europe.

This includes wind and solar projects that are either operational or will be soon. A large chunk of the power theyll produce will go toward offsetting the energy used by Apple customers. Use of everything from Mac Pros to Apple Watches accounts for nearly one-third of the companys carbon footprint.