Techcrunch:
Pony.ai and WeRide have received a key approval from Chinese securities regulators that clears the way for the autonomous vehicle technology companies to pursue secondary listings on the Stock Exchange of Hong Kong.

The Chinese companies, both of which are based in Guangzhou, are already publicly traded in the U.S. on the Nasdaq Exchange. WeRide made its public debut in October 2024, and Pony.ai followed a month later.