Engadget: Cybersecurity
PayPal's blockchain partner accidentally minted $300 trillion in stablecoins

Who wants to be a trillionaire?

Anna Washenko

Contributing Reporter

Anna WashenkoContributing Reporter

Wed, October 15, 2025 at 10:40 PM UTC

1 min read

MR.Cole_Photographer via Getty Images

In an embarrassing error, PayPal blockchain partner Paxos accidentally minted 300 trillion of the PYUSD token today. PayPal's stablecoin, which it debuted in 2023, is supposed to be redeemable at a 1:1 rate for US dollars. $300 trillion is more than double the entire world's current GDP of $117 trillion. In short, that's a really bad mistake from Paxos.

At least the action did appear to be a mistake. Paxos acknowledged the issue in a post on X, assuring clients that their funds are secure and stating that it burned the excess reserves of stablecoins. The company claimed that "this was an internal technical error.” Maybe so, but internal technical errors can still wreak havoc pretty quickly when it comes to finance, especially in the fantasy worlds of blockchain and crypto. Thanks to the appropriately named Web3 is Going Just Great for picking up on this story.

Advertisement

About our ads