Techcrunch:
The sprawling metropolis of Phoenix continues to be a testbed for Waymo, and not just on the technical front. The Alphabet-owned company has used its first testing and commercial market to experiment with its business model, too, including delivery.

Waymo is known for its robotaxis, but it has tested how its tech could be used for delivery. The company previously tapped UPS and Uber Eats for pilot programs, and it put considerable staff and resources toward developing self-driving trucks under its Waymo Via arm, before shuttering that program in 2023 to focus on robotaxis.