Techcrunch:
Digital rights group the Electronic Frontier Foundation (EFF) filed a lawsuit on Thursday against the Trump administration over the governments alleged social media monitoring of people who are lawfully living in the United States.

The suit contends that the government is using AI and other methods to watch the social media posts of virtually every single non-citizen who is in the United States legally and on a valid visa, as well as many people who have obtained permanent resident status. The U.S. government is, according to these allegations, looking for posts that express views that the current government disfavors.