Engadget: The Crew 2 is now playable offline
It's a welcome addition from Ubisoft.

Anna Washenko

Contributing Reporter

Anna WashenkoContributing Reporter

Thu, October 16, 2025 at 9:16 PM UTC

1 min read

Ubisoft

The Crew 2 was updated today to include Hybrid Mode, adding an offline mode to the driving game. Online and offline modes are separate saves, so if you snag a sweet ride while playing offline, it won't be available the next time you join an online session. Players will have the option to re-export their online save to the offline one, but it will overwrite and erase all offline-only progress. Multiplayer content, user-generated content, LIVE Summits and Crew Credits purchases will not be available in the offline mode.

It's a bare-bones option, but the addition of offline mode is a welcome move from Ubisoft. The company's decision to delete The Crew from players' libraries after servers for the online game shut down sparked some big debates about ownership and preservation. One of the notable voices emerging from that conversation was the Stop Killing Games movement, which is pushing for EU legislation to ensure access to games even after their devs stop supporting a project.

Having a way to continue accessing The Crew 2 even if (or more likely when) the game loses online support was something the developer had promised, so it's nice to have Ubisoft follow through. "Whether you're looking to preserve your progression for the future or simply enjoy the freedom of playing without a connection, Hybrid Mode ensures The Crew 2 remains accessible for years to come," the company said in the blog post.

Advertisement

About our ads