Engadget: Entertainment
TV & movies

Apple TV and Peacock team up on a bundle that costs $15 per month

That's a savings of around $9 every 30 days.

Lawrence Bonk

Contributing Reporter

Lawrence BonkContributing Reporter

Thu, October 16, 2025 at 5:04 PM UTC

1 min read

Apple/NBC

There's a new streaming bundle in town. Apple TV and Peacock are teaming up to offer a combined subscription that starts at $15 per month. This will allow access to all original programming from both platforms, in addition to Peacock's stable of network and cable shows.

This is a mighty fine deal, given that Apple TV recently changed both its name and price. The platform costs $13 per month now on its own. Peacock starts at $11 per month. In other words, this is a discount of around $9 each month. In this economy, we'll take any savings we can get.

There's one caveat here. The $15 price tag is for Peacock with ads. The subscription shoots up to $20 per month for an ad-free version. However, a standalone subscription to ad-free Peacock is $17 on its own. This bundle becomes available on October 20. Additionally, Apple One subscribers will get a 35 percent discount on Peacock Premium Plus plans. It's always nice when two lonely corporations find friendship, isn't it?

For the uninitiated, Apple TV is the company's big-wig streaming platform. It's primarily known for sci-fi like Severance, For All Mankind and the upcoming Pluribus. The platform is also host to plenty of comedy, like The Studio, Shrinking and Ted Lasso.

Peacock is NBC's streaming service. It streams old-school network programming like The Office, Grimm and Superstore. The service features a stable of original programming like Poker Face, Twisted Metal and the underrated Mrs. Davis. The platform also recently premiered a little show called The Paper, which is a spinoff of The Office. Against all odds, this is actually a great little sitcom and a worthy successor to the original.

Advertisement

About our ads