Engadget: Entertainment
What to read this weekend: Near Flesh and the return of 30 Days of Night

Vampires are heading back to Barrow.

Cheyenne MacDonald

Weekend Editor

Cheyenne MacDonaldWeekend Editor

Sat, October 18, 2025 at 9:51 PM UTC

Here are some recently released titles to add to your reading list. This week, we read Near Flesh, a collection of short stories by the late Katherine Dunn, and 30 Days of Night: Falling Sun, the revival of the hit horror series.

MCD

Near Flesh

This posthumous collection of short fiction is as strange and gritty and beautiful and horrifying as fans of Geek Love author Katherine Dunn would expect, and a great introduction to her voice and perspective for readers new to her work. Its 19 stories give us everything from a society obsessed with reuniting people with their amputated body parts when they're buried, to a teenager who yearns for salvation brought by alien visitors and a woman who surrounds herself with sex robots. Near Flesh is a dark and sometimes humorous exploration of the human condition, focusing mostly on women getting through their day-to-day lives, faced with both the mundane and the extreme. It hits pretty hard.

$27 at Amazon

IDW

30 Days of Night: Falling Sun #1

I love the 30 Days of Night franchise, so I was super excited to discover this week that it's picking back up with Falling Sun, and this first issue did not disappoint. 30 Days of Night: Falling Sun is written by Rodney Barnes (Killadelphia), who worked with the series' original writer Steve Niles to bring the new story to life, and features art by Chris Shehan (House of Slaughter). In 30 Days of Night: Falling Sun, 20 years have passed since vampires descended upon Barrow. Teenager Jalen James arrives in the remote Alaskan town just ahead of the long polar night to stay with his uncle, looking for an escape from his life in LA. But unbeknownst to the townspeople, the vampires have begun stirring again, with vengeance on their minds and their sights set on Barrow.

The first issue is really setting the stage, reintroducing us to the vampires and giving us a sense of who is who in Barrow now (and who is still haunted by the violence of the past). I've got high hopes for where it'll take us.

$6 at Amazon

Advertisement

About our ads