Engadget: Cybersecurity
Google has killed Privacy Sandbox

Privacy Sandbox was supposed to replace third-party cookies in Chrome.

Mariella Moon

Contributing Reporter

Mariella MoonContributing Reporter

Sat, October 18, 2025 at 1:00 PM UTC

2 min read

hapabapa via Getty Images

Google's Privacy Sandbox is officially dead. In an update on the project's website, Google Vice President Anthony Chavez has announced that the company was sunsetting the remaining technologies developed for Sandbox due to their "low levels of adoption." A spokesperson has confirmed to AdWeek that Google isn't just killing those technologies, it's retiring the whole initiative altogether. "We will be continuing our work to improve privacy across Chrome, Android and the web, but moving away from the Privacy Sandbox branding," the spokesperson said. "We're grateful to everyone who contributed to this initiative, and will continue to collaborate with the industry to develop and advance platform technologies that help support a healthy and thriving web."

The company launched Privacy Sandbox in 2019 as a future replacement to third-party cookies. It's a set of open standards that are supposed to enable personalized ads without divulging identifying data. Over the years, Google's plans to deprecate third-party cookies got pushed back again and again due to a series of delays and regulatory hurdles. Specifically, both the UK's Competition and Markets Authority (CMA) and the US Department of Justice looked into the Privacy Sandbox out of concerns that it could harm smaller advertisers.

In 2024, Google ultimately decided not to kill third-party cookies in Chrome and instead chose to roll out "a new experience in Chrome that lets people make an informed choice that applies across their web browsing." Just this April, Google announced that it wasn't going to make any to changes to how third-party cookies work on the Chrome browser at all, and that it was going to "maintain [its] current approach to offering users third-party cookie choice in Chrome." At the time, the company said that it was going to keep the Privacy Sandbox initiative alive, but things have clearly changed since then. Chavez wrote in the latest update that Google will "continue to utilize learnings from the retired Privacy Sandbox technologies."

Advertisement

About our ads