Techcrunch:
Home goods company Kohler recently unveiled a new device called the Dekoda — a $599 camera that can be attached to your toilet bowl and take pictures of whats inside.

CNET reports that the Dekoda analyzes these images in order to provide updates on your gut health and hydration, and to potentially detect blood. It also comes with a rechargeable battery, a USB connection, and a fingerprint sensor to identify whos using the toilet.