Techcrunch:
Welcome back to TechCrunch Mobility — your central hub for news and insights on the future of transportation. Want another reason to get this free in your inbox? The emailed version of this newsletter includes polls, including one this week that asks readers what they think the best business model is for autonomous vehicle technology. Have an opinion about this? Email me your opinion at kirsten.korosec@techcrunch.com with the subject line &#8220;AV poll.&#8221;

OK, back to the show. Theres another twist in the road for lidar company Luminar. And yes, it includes some inside-the-boardroom intrigue.