Techcrunch:
The Trump administration is taking its social media attacks on Democratic opponents to an unlikely platform, with the White House and numerous other government agencies joining Bluesky on Friday.

In its first post, the new White House account uploaded a highlight reel of Trump footage and memes, along with the message, “Whats up, Bluesky? We thought you mightve missed some of our greatest hits, so we put this together for you.”