Engadget: X is testing a new way of opening links in posts to improve engagement
It comes in response to creators' complaints about the reach of posts containing links.

Cheyenne MacDonald

Weekend Editor

Cheyenne MacDonaldWeekend Editor

Sun, October 19, 2025 at 9:12 PM UTC

2 min read

X/Nikita Bier

Links have proven to be a tricky thing when it comes to social media engagement. On platforms like Threads, data suggests that people often don't click the links in posts, and creators on X have complained that posts including links don't get as much visibility. X is now trying to solve that. Head of product Nikita Bier shared that the platform is "testing a new link experience" that lets people open a link without leaving the original post entirely, so they can still view the Like, Repost and other buttons. The test is rolling out to iOS users first.

We're testing a new link experience, starting on iOS -- to make it easier for your followers to engage with your post while browsing links.

For creators, a common complaint is that posts with links tend to get lower reach. This is because the web browser covers the post and… pic.twitter.com/oWraLpPwji

— Nikita Bier (@nikitabier) October 19, 2025

"To help get better signal, posts will now collapse to the bottom of the page so people can react while you're reading," Bier wrote alongside a screen recording of how it will work. As to why posts with links typically don't perform as well, Bier said, "This is because the web browser covers the post and people forget to Like or Reply. So X doesn't get a clear signal whether the content is any good."

Of course, X has been accused of throttling links to specific sites, and Elon Musk has said in the past that links "don't get as much attention" on the platform, so tweaking the popups will only go so far in addressing the reach issue. Bier added that posts should always "stand alone as great content so write a solid caption."

Advertisement

About our ads