Techcrunch:
Apple has rolled out a new feature that makes it easier to customize Liquid Glass to your liking. Now users will be able to set the appearance of Liquid Glass to either Clear or Tinted, with the help of a new setting included in the latest beta update of its mobile and desktop operating systems.

The addition indicates Apple is listening closely to user feedback in these early days of the iOS 26, iPadOS 26, and macOS 26 public launches. Apples likes to push through more significant changes, but it will often offer a fallback option for those who have a hard time adjusting.