Engadget: Gaming
EA will shut down The Sims Mobile in January

In-app purchases have been switched off and the company will delist the game on October 21.

Kris Holt

Contributing Reporter

Kris HoltContributing Reporter

Mon, October 20, 2025 at 5:39 PM UTC

2 min read

EA/Maxis

The Sims Mobile is going away as EA will shut it down on January 20, nearly eight years after its debut in March 2018. An update that went live on Monday is the final one and the company is delisting the game on October 21. Since the game requires online servers (game progress is stored there), it'll no longer be possible to play The Sims Mobile in three months' time.

"The Sims Mobile and The Sims community as a whole are full of creativity, kindness, and imagination," EA said in a blog post. "Youve amazed us with your stories, builds, and Sims. Were proud to have shared this journey with you, and as a token of our appreciation, weve prepared a few surprises as we head into TSMs final chapter."

EA will run 14 events, including limited-time quests and treasure hunts, in The Sims Mobile over the next few months. It's no longer possible to spend any real money in the game, though players can use their existing in-game currency. As of today, players also have unlimited energy, which "should make it easier to complete projects and enjoy the last few months of play," EA said. On January 6, EA will unlock all of the Build Mode and Create A Sim items so players can fully customize their homes and sims' looks for a couple of weeks before the grim reaper shows up for The Sims Mobile.

Advertisement

Advertisement

Advertisement

Back in 2022, EA and Sims studio Maxis announced Project Rene, a game that will run across multiple platforms, including mobile. So it's maybe not too surprising that the company wants to focus its Sims mobile efforts there.

In any case, the shutdown stinks, especially for those who have spent money on The Sims Mobile as well as for the sake of game preservation. The decision to scuttle the game won't exactly allay concerns that fans of the series, which has a significant LGBTQ+ fanbase, have about the future of The Sims if and when the planned sale of EA to a group that includes Saudi Arabia's Public Investment Fund goes through.

Advertisement

About our ads