Techcrunch:
As the entertainment industry reckons with when and how to use generative AI in filmmaking, Netflix is leaning in. In its quarterly earnings report released on Tuesday afternoon, Netflix wrote in its letter to investors that it is &#8220;very well positioned to effectively leverage ongoing advances in AI.&#8221;

Netflix isnt planning to use generative AI as the backbone of its content but believes the technology has potential as a tool to make creatives more efficient.