Techcrunch:
Ive often joked that I would love to have a pet if only animals did not need to poop and eat smelly, wet mush from a can. I want a fuzzy pal to hang out with all day, but then Ill hear that my friend spent $500 at the vet because their cat nibbled on a leaf, and the illusion breaks.

Its hard enough to take care of myself — do I really want to be responsible for a creature who might wake me up at 4 a.m. to pee?