Techcrunch:
Codi, an Andreessen Horowitz-backed startup founded by Christelle Rohaut and Dave Schuman, is launching what it hails as the first AI-powered platform to fully automate office management.

Codi was founded in 2018, in a pre-pandemic world, with a mission to help companies find flexible office spaces. It was more of a marketplace, as TechCrunch previously reported, that matched companies to buildings that offered flexible office arrangements. Codi then assisted with the move-in processes.