Engadget: Gaming
Xbox

Microsoft increases the price of Xbox dev kits by $500

The many price hikes announced during 2025 may burn the company in the end.

Anna Washenko

Contributing Reporter

Tue, October 21, 2025 at 7:00 PM UTC

Players aren't the only ones facing higher price tags from Xbox. According to a report by The Verge, Microsoft has upped the cost of the Xbox Development Kit from $1,500 to $2,000. That's a 33 percent jump in cost for these custom hardware kits, which are essential for devs to make and test games for release on the console.

"The adjustment reflects macroeconomic developments," Microsoft said in an email sent to Xbox devs and seen by The Verge. "We remain committed to providing high-quality tools and support for your development efforts." Although the macroeconomics in question are almost certainly the tariffs enacted by the US, it appears this is a blanket increase that will impact developers in other countries as well. The new kit costs appear to be effective immediately.

The change caps off a series of price increases for the Xbox ecosystem. Game Pass prices recently rose, with the Ultimate tier now costing $30 a month compared to the previous $20. And Microsoft has upped the cost of the Xbox twice this year, once in May and again in September. Between these additional expenses and the little matter of cutting thousands of gaming jobs earlier this year, a lot of us are giving up on Xbox before Microsoft can disappoint us yet again.

Advertisement

About our ads