Techcrunch:
Teslas record sales quarter has offered the company a reprieve after a terrible start to 2025. But CEO Elon Musk is focused on building a &#8220;robot army&#8221; and making good on his years-long, unfulfilled promise of self-driving cars &#8212; tasks he needs to accomplish if he is to unlock the full value of the $1 trillion compensation package that Tesla wants to award him.

The tension between Teslas current automotive-driven business and the AI-centric one that Musk is aiming for has never been more clear.