Techcrunch:
OpenAI unveiled its AI browser, ChatGPT Atlas, during a livestream on Tuesday. There are other AI browsers such as The Browser Companys Dia, Operas Neon, Perplexitys Comet, and General Catalyst-backed Strawberry. OpenAIs launch is notable because of the sheer scale of reaching potentially 800 million of its weekly ChatGPT users. For the company, the browser is much more about keeping ChatGPT central than about making web browsing better.

While Atlas is currently available only on the Mac, the company is already working on bringing it to Windows, iOS, and Android — all the surfaces where ChatGPT already exists. OpenAI has also made the browser available to all users instead of opting for an invite system like its rivals. The core proposition of the browser is for you to think of ChatGPT as the first interaction surface for search and answers instead of Google.