Techcrunch:
Also, the micromobility startup spun out of Rivian, has landed a commercial deal with Amazon to supply the e-commerce company with thousands of its new pedal-assist cargo quad vehicles that are big enough to carry more than 400 pounds of packages and small enough to use a bike lane.

Under the multi-year collaboration, the two companies will work on customizing the pedal-assist vehicles to meet Amazons delivery needs in Europe and the United States. The TM-Q pedal-assist electric quads will launch in spring 2026, according to Also, which was revealed on Wednesday at an event in Oakland alongside the companys new e-bike called the TM-B.