Engadget: Transportation
The first e-bike from Rivian spinoff Also has a virtual drivetrain

The startup also showed off two pedal-assisted electric quads it says are bike line compliant.

Karissa Bell

Senior Editor

UpdatedWed, October 22, 2025 at 9:29 PM UTC

The TM-B e-bike from Also.

(Also)

Ever since Rivian spun off its "micromobility business" into a standalone startup called Also earlier this year, there's been much speculation about what kind of vehicles the company is working on. Now, Also is showing off its first products: a lineup of e-bikes and two pedal-assisted electric quads.

The TM-B e-bike is Also's attempt at a do-it-all e-bike that can adapt to different use cases whether that's daily commuting, trail riding or kid and cargo-hauling. It sports a modular frame that can also accommodate a bench seat or rear cargo rack that supports up to 35KG of weight.

The different seats can be easily swapped out without extra tools. Instead, a button on the bikes touchscreen display controls a latching mechanism that releases the seat. It only comes in one frame size, but Also says it should be able to adapt to "multiple body sizes,” thanks to different seat sizes and styles.

The bench seat for the TM-B.

(Karissa Bell for Engadget)

The removable USB-C battery comes in two sizes: standard, which can power up to 60 miles of riding, and large, which maxes out at 100 miles of range. When youre not riding, the batteries can also be used as a large external battery pack.

Advertisement

Advertisement

Advertisement

In terms of power, the TM-Bs throttle tops out at 20MPH though the bike can reach speeds up to 28MPH with added pedaling. Also is taking an interesting approach to its drive system, with a setup it's labeled "DreamRide." Instead of a mechanical connection between the bike's rear wheel and the pedals, the TM-B uses "software-defined pedaling,"

In practice, this means that you pedaling is actually feeding the generator that powers the bikes battery rather than directly pushing you forward. However, an Also rep told me that there is also a “limp mode” for when the bike runs out of juice so riders wont get stranded. In those situations, pedaling will give the bike enough juice to hopefully get you to a spot where you can recharge.

Also has envisioned the TM-B in a lot of scenarios, many of which involve hauling a lot of cargo.

(Karissa Bell for Engadget)

Software-controlled pedaling probably wont appeal to purists, but Also says it enables a much more customizable riding experience. When in auto mode, the bike will adapt to the speed youre pedaling, though you can push on the throttle to get a boost. Theres also a manual mode that lets you select a “gear” (these are also software-controlled).

It also uses regenerative braking, so tapping on the brakes helps recharge the battery. Though in my short test ride I found that I didnt need to use the brakes much, because when I stopped pedaling the bike slowed down pretty quickly, kind of like taking your foot off the accelerator in an EV.

The Also app and Portal display.

(Also)

Given the bike's roots at Rivian, it's not surprising that there are also a bunch of other tech-enabled features, including a 5-inch touchscreen display, called "Portal," that supports navigation, music playback and calling features via an accompanying app. Theres also a built-in security system that automatically locks the frame and rear wheel when you walk away. On the handlebars, there are customizable controls that can be used to adjust the volume and music playback, answer calls or navigate through display.

Customizable controls on the left side fo the handlebar and a throttle on the right.

(Karissa Bell for Engadget)

Advertisement

Advertisement

Advertisement

Also is selling the TM-B in three configurations. The first to ship next spring will be the $4,500 TM-B Limited Launch Edition, which has a range up to 100 miles, support for standard and sport ride modes and features transparent purple accents. The $4,500 TM-B Performance has the same features as the limited edition model, but has a slightly different color scheme, and will be available within the "first half" of 2026. Finally, there's a base-level TM-B model with a range of up to 60 miles that only comes with standard ride modes. Also hasn't announced an exact price, but says it will cost less than $4,000 when it ships "later in 2026."

Pre-orders for the Launch Edition are open now and the other two bikes are available to reserve with a $50 deposit. The bikes will also be on display in Rivian showrooms later this year,

Also's quad for commercial uses cases (left) and a smaller quad for families (right).

(Also)

The company also previewed two electric, pedal-assisted quads it's calling TM-Q. The smaller quad is apparently meant for "families and individuals seeking a safe, compact alternative to cars" that can still haul “significant loads.” The larger TM-Q, on the other hand, is meant for commercial deliveries.

Also has partnered with Amazon to develop fleets of such vehicles that can be used by delivery drivers. Both quads are intended to be used in bike lanes, according to Also.

Also will partner with Amazon for a Prime-branded TM-Q.

(Karissa Bell for Engadget)

The company didn't share details about when these vehicles might be available or how much they'll cost.

Update, October 22, 2025, 2:29PM PT: Added more details and photos from Alsos launch event.

Advertisement

About our ads